+++ lib/rubocop/cop/mixin/minitest_exploration_helpers.rb
@@ -62,9 +62,16 @@ module RuboCop
         class_node.parent_class && class_node.identifier.source.end_with?('Test')
       end
 
+      def test_case?(node)
+        return false unless node.def_type? && test_case_name?(node.method_name)
+
+        class_ancestor = node.each_ancestor(:class).first
+        test_class?(class_ancestor)
+      end
+
       def test_cases(class_node)
         class_def_nodes(class_node)
-          .select { |def_node| def_node.method_name.to_s.start_with?('test_') }
+          .select { |def_node| test_case_name?(def_node.method_name) }
       end
 
       def lifecycle_hooks(class_node)
@@ -74,7 +81,6 @@ module RuboCop
 
       def class_def_nodes(class_node)
         class_def = class_node.body
-
         return [] unless class_def
 
         if class_def.def_type?
@@ -82,6 +88,24 @@ module RuboCop
         else
           class_def.each_child_node(:def).to_a
         end
+      end
+
+      def test_case_name?(name)
+        name.to_s.start_with?('test_')
+      end
+
+      def assertions(def_node)
+        method_def = def_node.body
+        return [] unless method_def
+
+        send_nodes =
+          if method_def.send_type?
+            [method_def]
+          else
+            method_def.each_child_node(:send)
+          end
+
+        send_nodes.select { |send_node| assertion?(send_node) }
       end
 
       def assertion?(node)