+++ test/rubocop/cop/minitest/literal_as_actual_argument_test.rb
@@ -41,6 +41,25 @@ class LiteralAsActualArgumentTest < Minitest::Test
     RUBY
   end
 
+  def test_registers_offense_when_actual_is_hash_without_curly_braces
+    assert_offense(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          assert_equal(foo, key: :value)
+                       ^^^^^^^^^^^^^^^^ Replace the literal with the first argument.
+        end
+      end
+    RUBY
+
+    assert_correction(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          assert_equal({key: :value}, foo)
+        end
+      end
+    RUBY
+  end
+
   def test_does_not_register_offense_when_actual_is_not_literal
     assert_no_offenses(<<~RUBY)
       class FooTest < Minitest::Test
@@ -56,6 +75,26 @@ class LiteralAsActualArgumentTest < Minitest::Test
       class FooTest < Minitest::Test
         def test_do_something
           assert_equal(2, foo)
+        end
+      end
+    RUBY
+  end
+
+  def test_does_not_register_offense_when_parens_omitted
+    assert_no_offenses(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          assert_equal 2, foo
+        end
+      end
+    RUBY
+  end
+
+  def test_does_not_register_offense_when_given_splat
+    assert_no_offenses(<<~RUBY)
+      class FooTest < Minitest::Test
+        def test_do_something
+          assert_equal(*foo)
         end
       end
     RUBY